docs: add StraightAnglePieChart example in TypeScript
